set multiple endpoints in config

+++ parser.h
@@ -1,6 +1,8 @@
 #ifndef PARSER_H
 #define PARSER_H
 
-const char *parse_request(char *buffer, long cnt, char **responses);
+#include <stddef.h>
+
+const char *parse_request(char *buffer, size_t cnt, char **responses);
 
 #endif
